Schedular1./requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
openpyxl==3.0.5
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.4
schedule==1.0.0
six==1.15.0
toml==0.10.2
